+++ Net/include/Poco/Net/Socket.h
@@ -306,22 +306,6 @@ protected:
 
 private:
 
-#if defined(POCO_HAVE_FD_POLL)
-class FDCompare
-	/// Utility functor used to compare socket file descriptors.
-	/// Used in poll() member function.
-{
-public:
-	FDCompare(int fd): _fd(fd) { }
-	inline bool operator()(const Socket& socket) const
-	{ return socket.sockfd() == _fd; }
-
-private:
-	FDCompare();
-	int _fd;
-};
-#endif
-
 	SocketImpl* _pImpl;
 };
