+++ Net/include/Poco/Net/SocketImpl.h
@@ -25,6 +25,7 @@
 #include "Poco/Net/SocketAddress.h"
 #include "Poco/RefCountedObject.h"
 #include "Poco/Timespan.h"
+#include <vector>
 
 
 namespace Poco {
@@ -46,6 +47,8 @@ public:
 		SELECT_WRITE = 2,
 		SELECT_ERROR = 4
 	};
+
+	typedef std::vector<SocketImpl*> SocketImplList;
 
 	virtual SocketImpl* acceptConnection(SocketAddress& clientAddr);
 		/// Get the next completed connection from the
@@ -171,6 +174,36 @@ public:
 	virtual int available();
 		/// Returns the number of bytes available that can be read
 		/// without causing the socket to block.
+
+	static int select(SocketImplList& readList, SocketImplList& writeList, SocketImplList& exceptList, const Poco::Timespan& timeout);
+		/// Determines the status of one or more sockets, 
+		/// using a call to select().
+		///
+		/// ReadList contains the list of sockets which should be
+		/// checked for readability.
+		///
+		/// WriteList contains the list of sockets which should be
+		/// checked for writeability.
+		///
+		/// ExceptList contains a list of sockets which should be
+		/// checked for a pending error.
+		///
+		/// Returns the number of sockets ready.
+		///
+		/// After return, 
+		///   * readList contains those sockets ready for reading,
+		///   * writeList contains those sockets ready for writing,
+		///   * exceptList contains those sockets with a pending error.
+		///
+		/// If the total number of sockets passed in readList, writeList and
+		/// exceptList is zero, select() will return immediately and the
+		/// return value will be 0.
+		///
+		/// If one of the sockets passed to select() is closed while
+		/// select() runs, select will return immediately. However,
+		/// the closed socket will not be included in any list.
+		/// In this case, the return value may be greater than the sum
+		/// of all sockets in all list.
 		
 	virtual bool poll(const Poco::Timespan& timeout, int mode);
 		/// Determines the status of the socket, using a 
@@ -417,6 +450,22 @@ protected:
 		/// Throws an appropriate exception for the given error code.
 
 private:
+#if defined(POCO_HAVE_FD_POLL)
+class FDCompare
+	/// Utility functor used to compare socket file descriptors.
+	/// Used in poll() member function.
+{
+public:
+	FDCompare(int fd): _fd(fd) { }
+	inline bool operator()(const SocketImpl* socketImpl) const
+	{ return socketImpl->sockfd() == _fd; }
+
+private:
+	FDCompare();
+	int _fd;
+};
+#endif
+
 	SocketImpl(const SocketImpl&);
 	SocketImpl& operator = (const SocketImpl&);
